+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Implement space padding on heading separation / justification row.
 - Added unit testing.
 - Clarify descriptions of functionality based on new target of matching markdownlint/MD060.
+- Empty placeholder empty by default.
 
 ## 1.0.0